Changed RTCTimer to Timer and moved into Wiring
