+++ hal/src/bluz/apptimer_hal.c
@@ -1,4 +1,4 @@
-#include "apptimer_hal.h"
+#include "app_timer.h"
 #include <nordic_common.h> // MAX
 
 uint32_t HAL_app_timer_create(app_timer_id_t *p_timer_id, app_timer_mode_t mode, app_timer_timeout_handler_t timeout_handler)
