--- platform/MCU/NRF51/SPARK_Firmware_Driver/src/rtc_timer.cpp
@@ -1,2 +1,0 @@
-#include "rtc_timer.h"
-
